requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.6
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
